add pagination and connect contacts with owners

+++ model/contacts.js
@@ -1,27 +1,46 @@
 const Contacts = require('./schemas/contactsSchema')
 
-const listContacts = async () => {
-  return await Contacts.find()
+const listContacts = async (userId, query) => {
+  const { favorite, page = 1, limit = 10 } = query
+  const options = { owner: userId }
+  if (favorite) options.favorite = favorite
+  return await Contacts.paginate(options, {
+    page,
+    limit,
+    populate: {
+      path: 'owner',
+      select: 'email subscription'
+    }
+  })
 }
 
-const getContactById = async (contactId) => {
-  return Contacts.findById(contactId)
+const getContactById = async (userId, contactId) => {
+  return Contacts.findById({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription'
+  })
 }
 
-const removeContact = async (contactId) => {
-  return await Contacts.findByIdAndDelete(contactId)
+const removeContact = async (userId, contactId) => {
+  return await Contacts.findByIdAndDelete({ _id: contactId, owner: userId })
 }
 
-const addContact = async (body) => {
-  return await Contacts.create({ ...body })
+const addContact = async (userId, body) => {
+  return await Contacts.create({ ...body, owner: userId })
 }
 
-const updateContact = async (contactId, body) => {
-  return  await Contacts.findByIdAndUpdate(contactId, body, { new: true })
+const updateContact = async (userId, contactId, body) => {
+  return  await Contacts.findByIdAndUpdate({ _id: contactId, owner: userId }, body, { new: true }).populate({
+    path: 'owner',
+    select: 'email subscription'
+  })
 }
 
-const updateStatusContact = async (contactId, body) => {
-  return  await Contacts.findByIdAndUpdate(contactId, body, { new: true })
+const updateStatusContact = async (userId, contactId, body) => {
+  return  await Contacts.findByIdAndUpdate({ _id: contactId, owner: userId }, body, { new: true }).populate({
+    path: 'owner',
+    select: 'email subscription'
+  })
 }
 
 module.exports = {
